updated test for loop. all test are inside switch statement in main.

+++ Test.cpp
@@ -11,7 +11,7 @@ void LD_register_immediate_test(Bus& test) {
 	test.reset();
 
 	// load LD r,n test instructions into ram
-	for (uint16_t i = 0; i < memory.size(); i++) {
+	for (int i = 0; i < memory.size(); i++) {
 		test.ram[i] = memory[i];
 		std::cout << "ram[" << std::setfill('0') << std::setw(4) << std::hex << std::uppercase << i << "] : " <<
 			std::setw(2) << std::setfill('0') << static_cast<int>(test.ram[i]) << '\n';
@@ -40,7 +40,7 @@ void LD_register_register_test(Bus& test) {
 	// clear RAM
 	test.reset();
 
-	for (uint16_t i = 0; i < memory.size(); i++) {
+	for (int i = 0; i < memory.size(); i++) {
 		test.ram[i] = memory[i];
 		std::cout << "ram[" << std::setfill('0') << std::setw(4) << std::hex << std::uppercase << i << "] : " <<
 			std::setw(2) << std::setfill('0') << static_cast<int>(test.ram[i]) << '\n';
@@ -64,7 +64,7 @@ void LD_register_register_indirect_test(Bus& test) {
 	test.reset();
 
 
-	for (uint16_t i = 0; i < memory.size(); i++) {
+	for (int i = 0; i < memory.size(); i++) {
 		test.ram[i] = memory[i];
 		std::cout << "ram[" << std::setfill('0') << std::setw(4) << std::hex << std::uppercase << i << "] : " <<
 			std::setw(2) << std::setfill('0') << static_cast<int>(test.ram[i]) << '\n';
@@ -84,7 +84,7 @@ void LD_register_implied_test(Bus& test) {
 	test.reset();
 
 
-	for (uint16_t i = 0; i < memory.size(); i++) {
+	for (int i = 0; i < memory.size(); i++) {
 		test.ram[i] = memory[i];
 		std::cout << "ram[" << std::setfill('0') << std::setw(4) << std::hex << std::uppercase << i << "] : " <<
 			std::setw(2) << std::setfill('0') << static_cast<int>(test.ram[i]) << '\n';
@@ -110,7 +110,7 @@ void LD_register_indexed_test(Bus& test) {
 
 	test.reset();
 
-	for (uint16_t i = 0; i < memory.size(); i++) {
+	for (int i = 0; i < memory.size(); i++) {
 		test.ram[i] = memory[i];
 		std::cout << "ram[" << std::setfill('0') << std::setw(4) << std::hex << std::uppercase << i << "] : " <<
 			std::setw(2) << std::setfill('0') << static_cast<int>(test.ram[i]) << '\n';
@@ -129,7 +129,7 @@ void LD_register_extended_test(Bus& test) {
 
 	test.reset();
 
-	for (uint16_t i = 0; i < memory.size(); i++) {
+	for (int i = 0; i < memory.size(); i++) {
 		test.ram[i] = memory[i];
 		std::cout << "ram[" << std::setfill('0') << std::setw(4) << std::hex << std::uppercase << i << "] : " <<
 			std::setw(2) << std::setfill('0') << static_cast<int>(test.ram[i]) << '\n';
@@ -148,11 +148,11 @@ void LD_register_indirect_immediate(Bus& test) {
 
 	test.reset();
 
-	for (uint16_t i = 0; i < memory.size(); i++) {
-		test.ram[i] = memory[i];
-		std::cout << "ram[" << std::setfill('0') << std::setw(4) << std::hex << std::uppercase << i << "] : " <<
-			std::setw(2) << std::setfill('0') << static_cast<int>(test.ram[i]) << '\n';
-	}
+    for (int i = 0; i < memory.size(); i++) {
+        test.ram[i] = memory[i];
+        std::cout << "ram[" << std::setfill('0') << std::setw(4) << std::hex << std::uppercase << i << "] : " <<
+                  std::setw(2) << std::setfill('0') << static_cast<int>(test.ram[i]) << '\n';
+    }
 
 	for (int i = 0; i < 3; i++) {
 		test.cpu.instruction_cycle();
